backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pandas==2.1.4
python-dotenv==1.0.1
requests==2.31.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
python-multipart==0.0.9
aiohttp==3.9.3
sqlalchemy==2.0.27
jinja2==3.1.3
aiofiles==23.2.1 